added better UI tests
