+++ hackathon_site/dashboard/frontend/src/components/cart/CartCard/CartCard.tsx
@@ -124,7 +124,7 @@ export const CartCard = ({ hardware_id, quantity, error }: CartCardProps) => {
                     onClick={() => {
                         handleRemove(hardware_id);
                     }}
-                    role="remove"
+                    data-testid={hardware_id.toString()}
                 >
                     <CloseIcon />
                 </IconButton>
