+++ hackathon_site/dashboard/frontend/src/slices/hardware/cartSlice.ts
@@ -23,7 +23,7 @@ const cartAdapter = createEntityAdapter<CartItem>({
 
 export const cartReducerName = "cart";
 export const initialState = cartAdapter.getInitialState(extraState);
-export type CartItemState = typeof initialState;
+export type CartState = typeof initialState;
 
 // Slice
 const cartSlice = createSlice({
